setup user and jwt

+++ services/link.go
@@ -7,18 +7,18 @@ import (
 	"github.com/amirfakhrullah/go-bitly/model"
 )
 
-func GetAllLinks() ([]model.Link, error) {
+func GetAllLinks(userId uint) ([]model.Link, error) {
 	var links []model.Link
 
-	if err := db.DB.Find(&links).Error; err != nil {
+	if err := db.DB.Find(&links, "user_id = ?", userId).Error; err != nil {
 		return []model.Link{}, err
 	}
 	return links, nil
 }
 
-func GetLink(id uint64) (model.Link, error) {
+func GetLink(id uint64, userId uint) (model.Link, error) {
 	var link model.Link
-	if err := db.DB.Find(&link, "id = ?", id).Error; err != nil {
+	if err := db.DB.Find(&link, "id = ? AND user_id = ?", id, userId).Error; err != nil {
 		return model.Link{}, err
 	}
 	if link.ID == 0 {
@@ -32,9 +32,9 @@ func CreateLink(l model.Link) error {
 	return tx.Error
 }
 
-func UpdateLink(updatedLink model.Link, id uint64) error {
+func UpdateLink(userId uint, updatedLink model.Link, id uint64) error {
 	var link model.Link
-	db.DB.Find(&link, "id = ?", id)
+	db.DB.Find(&link, "id = ? AND user_id = ?", id, userId)
 	link.RedirectUrl = updatedLink.RedirectUrl
 	link.ShortenedId = updatedLink.ShortenedId
 	link.Clicked = updatedLink.Clicked
@@ -42,13 +42,18 @@ func UpdateLink(updatedLink model.Link, id uint64) error {
 	return tx.Error
 }
 
-func DeleteLink(id uint64) error {
-	tx := db.DB.Delete(&model.Link{}, "id = ?", id)
+func DeleteLink(userId uint, id uint64) error {
+	tx := db.DB.Delete(&model.Link{}, "id = ? AND user_id = ?", id, userId)
 	return tx.Error
 }
 
-func FindByShortenedId(id string) (model.Link, error) {
+// PUBLIC --------------------
+func OpenShortenedId(id string) (model.Link, error) {
 	var l model.Link
-	tx := db.DB.Find(&l, "shortened_id = ?", id)
+	if err := db.DB.Find(&l, "shortened_id = ?", id).Error; err != nil {
+		return model.Link{}, err
+	}
+	l.Clicked += 1
+	tx := db.DB.Save(&l)
 	return l, tx.Error
 }
